requirements.txt
click==8.1.7
colorama==0.4.6
joblib==1.4.0
lxml==5.2.1
nltk==3.8.1
python-docx==1.1.0
regex==2024.4.16
tqdm==4.66.2
typing_extensions==4.11.0
